fix: rename help entity file
