+++ src/game/entities/help-entity.ts
